Use canonical order of rooms as exit options
Previously the exit options would be sorted by the `room` object's
key iteration order. That iteration order is arbitrary
(https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/for...in)
so for games with lots of rooms, it can be really inconvenient to
have to search though an unpredictable list of room names for the
one you want. This instead uses the order of rooms that correspond
to moving forward/backward through the room selector card.