fix: rainbow text effect letters not showing up when dialog completed early
because the logic here is `(time - position) % count`, with `time` potentially be zero and `position` always being positive, you can end up with e.g. `-11 % 10`, which in JS results in `-1`, which is an invalid palette index. an alternate modulo fn is added which will ensure a positive number following the expected pattern (e.g. `positiveModule(-11, 10)` will give `9`)