Fix exported games breaking due to CRLF in template
Commit 7020c5c pushed a new updated packed version of the resources
files using editor/script/generated/resources.js. This script copies
Unicode data from files directly into a JS object. However, git will
check out files differently based on whether the user is on a Windows
machine, adjusting the newline format of text files automatically. When
the script was run on a Windows machine, it inserted CRLF newlines
instead of LF newlines into every file, including the template for
exported games. The games themselves worked fine when run in a browser,
but the editor's import tool depended on specific strings (including
newline characters) to find the game data.

This tool patches both the import tool and the resource packer to
replace CRLF newlines with regular LF ones. The import tool had to be
patched in order to load any games exported with Bitsy 7.11, and the
packer was changed in order to prevent any future problems.