revert "make sure ClearCache() deletes tile memory"
fixes #250

my ClearCache change broke the editor; the reset-room code
that gets called when edits are made is too complex for me
to reason about. I added some logging to ClearCache (after making this
commit) and noticed that ClearCache is called 4 separate times
every time you make an edit in the sprite Paint tool -- seems odd

stepping back, I think that the room should not reset itself on
every stroke of the Paint tool. but I'd rather not break anything
else so I'm not going to dig into it on my own.

anyway, my bad! this commit undoes my previous commit, which fixes #250

but now #230 is back, and I don't know how to fix it