version 8.12 (#265)
* Add regenerate button for blips (#236)

* Regenerate blips without creating a new one
* Regenerating a blip does not modify its name

* Write Guide for lockedDoor.md (#257)

* Write Guide for lockedDoor.md

Write step by step guide with images explaining how to create a locked door

* Fix build crash for lockedDoor.md

The format for one of the images caused the build to crash. Attempt to fix it.

* Fix another build crash in lockedDoor.md

The crash seems to be related to the images placed side-by-side, but I'm not entirely sure. Hopefully this fixes it.

* Add new image in lockedDoor.md

I combined the two separate images, so hopefully npm doesn't get upset and crash this time.

* Remove image in lockedDoor.md

I'm not sure why the last image kept causing a crash. Removing it for now, and maybe someone can help me add it back later.

* Add image back to lockedDoor.md

npm crashed even without the image, so that was not the problem. Trying to rewrite the bottom of the file in case there was a hidden character causing the crash.

* Removing Step 5 in lockedDoor.md

Seeing if npm crashes without the last step written.

* Update lockedDoor.md

I'm not sure what's causing the npm crash. I'm just going to leave the file so it looks how I intended in the preview, and hopefully someone can help me fix the npm crash.

* Add closing tags to img in lockedDoor.md

* v8.12 updates

---------

Co-authored-by: Alex Nash <contact@alexnash.dev>
Co-authored-by: vaasut <vaasut1@gmail.com>